hardlink.py: Ignore symbolic links
By using os.lstat instead of os.stat, we can check if the
file is a symbolic link or not. Because all non-regular files
are excluded, symlinks are excluded aswell.

This removes many OSError messages when the target of a
symbolic link can not be found. Symbolic links are another
form of link and should not be converted to hardlinks.