hardlink.py: Implement __lt__ for file, select best candidate as master
The File class gains a new __lt__() method which respects options given
on the commandline.

The behavior has been changed to select the file as master which is the
best in relation to the given options; ie. if the link count should be
maximized, the file with the highest link count becomes master.

This also changes the behavior of -t to retain the newest timestamp by
default, which makes everything more predictable.

Credits to Joachim Breitner <mail@joachim-breitner.de> for providing a patch
which inspired this patch.