Add file basename to files_by_ino, if --respect-name is specified
Implementing --respect-name in combination with grouping by inode
numbers is not nice, but this solution looks by far the easiest.

Another option is to keep the files grouped only by device and
inode, and then to handle file names in the linking code. This
would reduce the number of file objects we need to create, and
potentially reduce the number of comparisons, but is more
complicated code-wise.