hardlink: Handle extended attributes.
This change allows the user to specify that extended attributes are
to be taken into account when hardlink looks for "equal" files to link
together.

The new command-line option --respect-xattrs (or -X) means that for
any pair of files to be considered "equal", they must have exactly the
same set of attribute names, and exactly the same values assigned to
each of those corresponding attributes (in addition to all other
criteria).

Xattr equality is the penultimate criterion that is tested for a given
pair of files, just before the final test for content equality, since
those two are usually the most expensive criteria to evaluate and
xattr equality is usually cheaper than content equality.

The implementation avoids O(N^2) time by sorting the two sets of
attribue names (N*log(N)), before testing their equality and the
equality of thier values (O(N)); and it always bails out as early as
possible (e.g. even before sorting if the total lengths of the xattr
names are not equal).

This also includes a test script in test/test_hardlink.sh which
creates a bunch of files with various combinations of xaddrs, lists
them with their inode number, then runs hardlink on them, and then
lists them with inodes again. Since name of each file corresponds to
is xattrs, you can easily see if expected linking is indeed happening.

Issue: STPK-1902
Change-Id: I0a24924c7f5ab55cd9f4c7e0e24df1b127ef50c0
Signed-off-by: Tom Keel <thomas.keel@intel.com>