hardlink: Add --keep-oldest option.
This change introduces the --keep-oldest (-O) hardlink command-line
option. Specifying that option causes hardlink to keep the oldest
(mtime) of several inodes that have the same content. Prior to this
change the newest inode was kept, and that is still the default.

Note that if --minimize or --maximize is specified, that has a higher
precedence than --keep-oldest (or its absence).

Issue: STPK-1902
Change-Id: I1966d5ed1849afa4362923b95e42f2ef4be0cd7b
Signed-off-by: Tom Keel <thomas.keel@intel.com>